Issue #26239 refactor packages per spec updates
