+++ dev/com.ibm.websphere.appserver.features/visibility/public/data-1.0/io.openliberty.data-1.0.feature
@@ -7,6 +7,7 @@ IBM-API-Package: \
   io.openliberty.data.repository; type="ibm-api",\
   jakarta.data; type="spec",\
   jakarta.data.exceptions; type="spec",\
+  jakarta.data.page; type="spec",\
   jakarta.data.repository; type="spec"
 Subsystem-Name: Jakarta Data 1.0
 #TODO io.openliberty.jakartaeePlatform-11.0
