+++ dev/io.openliberty.data.internal.persistence/src/io/openliberty/data/internal/persistence/PaginatedIterator.java
@@ -22,7 +22,7 @@ import com.ibm.websphere.ras.TraceComponent;
 import com.ibm.websphere.ras.annotation.Trivial;
 import com.ibm.ws.ffdc.annotation.FFDCIgnore;
 
-import jakarta.data.repository.Pageable;
+import jakarta.data.page.Pageable;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.TypedQuery;
