+++ dev/io.openliberty.data.internal_fat/test-applications/ProviderTestApp/src/test/jakarta/data/inmemory/web/Composites.java
@@ -5,8 +5,8 @@ package test.jakarta.data.inmemory.web;
 
 import java.util.List;
 
+import jakarta.data.Limit;
 import jakarta.data.repository.DataRepository;
-import jakarta.data.repository.Limit;
 import jakarta.data.repository.Repository;
 
 /**
