+++ dev/io.openliberty.data.internal_fat_jpa/test-applications/DataJPATestApp/src/test/jakarta/data/jpa/web/CityAttrNames2.java
@@ -10,7 +10,7 @@
  *******************************************************************************/
 package test.jakarta.data.jpa.web;
 
-import jakarta.data.repository.StaticMetamodel;
+import jakarta.data.StaticMetamodel;
 
 /**
  * Static metamodel for the City entity.
