+++ dev/io.openliberty.jakarta.data.1.0/bnd.bnd
@@ -25,7 +25,7 @@ Bundle-Description: Copied Jakarta Data API classes and some proposed or experim
 Export-Package: \
   jakarta.data;version="1.0.0",\
   jakarta.data.exceptions;version="1.0.0",\
-  jakarta.data.provider;version="1.0.0",\
+  jakarta.data.page;version="1.0.0",\
   jakarta.data.repository;version="1.0.0"
 
 instrument.classesExcludes: jakarta/data/*.class
