+++ dev/io.openliberty.jakarta.data.1.0/src/jakarta/data/Direction.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2022,2023 IBM Corporation and others.
+ * Copyright (c) 2022 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -10,18 +10,11 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package jakarta.data.repository;
-
-import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
+package jakarta.data;
 
 /**
- * Method signatures copied from the Jakarta Data git repo.
+ * Copied from the jakarta.data.repository.Direction from the Jakarta Data repo.
  */
-@FunctionalInterface
-public interface Streamable<T> extends Iterable<T> {
-    public default Stream<T> stream() {
-        return StreamSupport.stream(this.spliterator(),
-                                    false /* not parallel */);
-    }
+public enum Direction {
+    ASC, DESC
 }
