+++ dev/io.openliberty.jakarta.data.1.0/src/jakarta/data/page/Page.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2022 IBM Corporation and others.
+ * Copyright (c) 2022,2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -10,9 +10,14 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package jakarta.data.repository;
+package jakarta.data.page;
 
 /**
+ * Method signatures copied from proposals in the Jakarta Data git repo.
  */
-public interface KeysetAwarePage<T> extends KeysetAwareSlice<T>, Page<T> {
+public interface Page<T> extends Slice<T> {
+
+    long totalElements();
+
+    long totalPages();
 }
