+++ dev/io.openliberty.jakarta.data.1.0/src/jakarta/data/repository/PageableRepository.java
@@ -12,6 +12,9 @@
  *******************************************************************************/
 package jakarta.data.repository;
 
+import jakarta.data.page.Page;
+import jakarta.data.page.Pageable;
+
 /**
  * Interface methods copied from the Jakarta Data git repository.
  */
